review: address second round of comments
- comments and naming changes from the PR review.
- RunbookExecutionMode, clone-address, and watchdog tests move to
  simnet/startup.rs with their subjects.
